test : LottoIssuerTest 제거 및 WinningLottoTest 에 WinningLotto 생성 테스트 추가
