+++ src/test/java/model/WinningLottoTest.java
@@ -1,6 +1,7 @@
 package model;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -54,6 +55,23 @@ class WinningLottoTest {
         assertThat(winningLotto).isInstanceOf(WinningLotto.class);
         assertThat(winningLotto.getLotto()).isEqualTo(lotto);
         assertThat(winningLotto.getBonusNo()).isEqualTo(bonusLottoNumber);
+    }
+
+    @DisplayName("입력받은 값으로 Lotto 생성을 테스트 한다.")
+    @Test
+    void issueWinningLotto() {
+        // given
+        String inputWinningLottoNumbers = "1,2,3,4,5,6";
+        String inputBonusNumber = "7";
+        Lotto inputWinningLotto = Lotto.generateLottoByInput(inputWinningLottoNumbers);
+        LottoNumber bonusNumber = new LottoNumber(Integer.valueOf(inputBonusNumber));
+
+        // when
+        WinningLotto winningLotto = WinningLotto.issueWinningLotto(inputWinningLotto, bonusNumber);
+
+        // then
+        assertThat(winningLotto.getLotto()).isEqualTo(inputWinningLotto);
+        assertThat(winningLotto.getBonusNo()).isEqualTo(bonusNumber);
     }
 
 }
